requirements.txt
aiogram==3.10.0
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
redis==5.0.1
yandex-cloud
